refactor: Move User DTOs to user package
- Moved UserFindByQueryResponseDTO and UserCreateRequestDTO to the com.leaf.domain.dtos.user package.
- Updated package declarations in both DTO files to reflect the new package structure.

This change organizes the DTOs into a user-specific package, improving the project structure and maintainability.
